+++ bluemira/base/constants.py
@@ -204,6 +204,48 @@ def gcm3_to_kgm3(density: Union[float, np.array, List[float]]) -> Union[float, n
     return density * 1000.0
 
 
+def pam3s_to_mols(flow_in_pam3_s):
+    """
+    Convert a flow in Pa.m^3/s to a flow in mols.
+
+    Parameters
+    ----------
+    flow_in_pam3_s: Union[float, np.array]
+        The flow in Pa.m^3/s to convert
+
+    Returns
+    -------
+    flow_in_mols: Union[float, np.array]
+        The flow in mol/s
+
+    Notes
+    -----
+    At 273.15 K for a diatomic gas
+    """
+    return flow_in_pam3_s / 2270
+
+
+def mols_to_pam3s(flow_in_mols):  # noqa :N802
+    """
+    Convert a flow in Pa.m^3/s to a flow in mols.
+
+    Parameters
+    ----------
+    flow_in_mols: Union[float, np.array]
+        The flow in mol/s to convert
+
+    Returns
+    -------
+    flow_in_pam3_s: Union[float, np.array]
+        The flow in Pa.m^3/s
+
+    Notes
+    -----
+    At 273.15 K for a diatomic gas
+    """
+    return flow_in_mols * 2270
+
+
 # =============================================================================
 # Working constants
 # =============================================================================
